comment fixing and refactoring
