+++ src/components/molecules/VenueCard/index.ts
@@ -1,0 +1,1 @@
+export { VenueCard } from "./VenueCard";
