--- src/components/molecules/VenueHero/index.ts
@@ -1,1 +1,0 @@
-export { VenueHero } from "./VenueHero";
