+++ src/pages/Admin/Venue/Details/VenueDetails.styles.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { Container as VenueHeroContainer } from "components/molecules/VenueHero/VenueHero.styles";
+import { Container as VenueHeroContainer } from "components/molecules/VenueCard/VenueCard.styles";
 import { Button } from "components/atoms/Button/Button.styles";
 
 export const Container = styled.div`
