+++ src/scss/constants.scss
@@ -51,6 +51,10 @@ $room-info-bg: #6f43ff;
 
 $small-laptop-width: 980px;
 
+$small-desktop-width: 1024px;
+
+$normal-tablet-width: 800px;
+
 // @debt the following locations are using styled-components, so can't use our helper function yet
 //   pages/Admin/Room/Modal/RoomModal.styles.ts (z-index: 101)
 //   components/molecules/UserProfilePicture/UserProfilePicture.styles.ts (z-index: 10)
